Fixed some more bugs. Changed the file index to long. Updated hwo the internal index is generated. Added a lookup for entries by their internal id. Fixed the storing loop with a list-iterator. Added more tests. Added a lot of logs.

+++ SemanticDecomposition/src/main/java/de/dailab/nsm/decomposition/Dictionaries/customDictionary/CustomEntry.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
 
 public class CustomEntry implements Serializable {
     public int index;
-    public int fileIndex = -1;
+    public long fileIndex = -1;
     public String word = null;
     public WordType type = WordType.UNKNOWN;
